chore(session-file): show confirm modal when deleting file (#485)
Currently, in the application when a file is being deleted,
there exists no confirmation upon clicking the delete button
as to if a user is sure about deleting the file. This can
lead to mistakenly deleted files.

This commit ensures a confirmation modal is shown whenever
the delete button is clicked for a session file, if the user
signifies intention to delete on the confirmation modal, only
then will the session file be deleted.

[Finishes #157051707]

+++ src/app/modules/request-pool/session-details/session-details.component.ts
@@ -4,6 +4,7 @@ import {
 import * as moment from 'moment';
 import { UserService } from '../../../services/user.service';
 import { FileService } from '../../../services/files.service';
+import { AlertService } from '../../../services/alert.service';
 
 @Component({
   selector: 'app-session-details',
@@ -34,7 +35,8 @@ export class SessionDetailsComponent implements OnInit {
 
   constructor(
     private userService: UserService,
-    private fileService: FileService) { }
+    private fileService: FileService,
+    private alertService: AlertService) { }
 
   ngOnInit() {
     this.userIsMentor = (this.currentUserId === this.request.mentor.id);
@@ -147,11 +149,18 @@ export class SessionDetailsComponent implements OnInit {
    * @return {void}
    */
   deleteFile(session, file) {
-    this.fileService.deleteSessionFile(session.id, file.id)
-      .toPromise()
-      .then((response) => {
-        this.removeDeletedFile.emit(file);
-      });
+    this.alertService
+      .confirm(
+        `Are you sure want to delete ${file.name}?`,
+        this, {
+          confirmActionText: `DELETE`,
+          abortActionText: 'CANCEL',
+          confirmAction: () => {
+            this.fileService.deleteSessionFile(session.id, file.id)
+              .toPromise()
+              .then(() => this.removeDeletedFile.emit(file));
+          },
+        });
   }
 
   /** Opens the add file modal.
